requirements/common.txt
splines==0.3.0
plyfile==1.1
tensorboard
wandb
tqdm
viser==0.2.23
opencv-python-headless==4.10.*
open3d
matplotlib
mediapy==1.2.2
einops
scikit-learn
numpy==1.26
torchmetrics==1.6.3
jsonargparse[signatures]>=4.27.7
submodules/simple-knn
submodules/diff-gaussian-rasterization
submodules/diff-gaussian-rasterization_ms
submodules/diff-gaussian-rasterization_radegs
submodules/fused-ssim
submodules/nvdiffrast